+++ ext/kino/src/lib.rs
@@ -8,6 +8,7 @@ mod control;
 mod cpus;
 mod env_strings;
 mod gvl;
+mod listen;
 mod logsink;
 mod mono;
 mod pin;